irrev/pyproject.toml
[project]
name = "irrev"
version = "0.1.0"
description = "Semantic compiler for the irreversibility vault"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "python-frontmatter>=1.1.0",
    "rich>=14.2.0",
    "watchdog>=4.0.0",
    "pygls>=1.3.0",
    "lsprotocol>=2023.0.0",
]

[project.scripts]
irrev = "irrev.cli:cli"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
